reconfiguring Firebase Auth (yes again) to work with GH Pages
Firebase was the worst decision made about this project.  It is impossibly brittle, making small
changes to your hosting platform is devistating to how the library is configured and
requires implementation overhauls as that features are regularly breaking on the library.

If I ever do this again, I will setup a login with github and roll golang lambda to solve it.